+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 本文件格式遵循 [Keep a Changelog](https://keepachangelog.com/zh-CN/1.1.0/)，版本号遵循语义化版本。
 
-## [Unreleased]
+## [2.3.2] - 2026-09-04
 
 ### 修复